requirements.txt
fastapi
Pillow
pre-commit
pyheif
python-multipart
requests
uvicorn[standard]
whatimage